Dedepe stats view code. Match to original
