+++ srtla-lib/src/main/res/layout/srtla_stats_view.xml
@@ -17,8 +17,7 @@
     android:paddingRight="12dp"
     android:paddingBottom="8dp"
     android:background="#F0F0F0"
-    android:layout_marginBottom="4dp"
-    android:visibility="gone" />
+    android:layout_marginBottom="4dp" />
 
   <!-- Container for dynamically added connection items -->
   <LinearLayout
@@ -28,6 +27,7 @@
     android:orientation="vertical"
     android:layout_marginBottom="16dp">
 
+    <!-- Default "no connections" message -->
     <TextView
       android:id="@+id/text_no_connections"
       android:layout_width="match_parent"
